refactor(structure): hymns + invitations into hooks/utils/__tests__ layout
- hymns: hymns.ts + hymns.json → utils/
- invitations:
  * 10 hooks → hooks/ (joining useLongPress, useMinuteTick from prior PR)
  * 10 utils → utils/
  * 3 util tests → utils/__tests__/
  * twilioClientProvider.tsx → TwilioChatProvider.tsx (PascalCase fix
    to match the exported component name)
  * InvitationLinkActions promoted to a folder because it has tests:
    {InvitationLinkActions.tsx, index.tsx, __tests__/}
  * vi.mock path in InvitationLinkActions.test.tsx updated to match
    its new relative location (../../utils/invitationsCallable)
- 14 external importers updated for the rename + path moves

Components stay flat. Per skill rule: only InvitationLinkActions earned
a folder (it has a test).

Typecheck + 346 tests + build all pass.
