+++ src/app/routes/invite-speaker.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import { useParams } from "react-router";
 import { useFullViewportLayout } from "@/hooks/useFullViewportLayout";
 import { SpeakerInvitationChat } from "@/features/invitations/SpeakerInvitationChat";
-import { TwilioChatProvider } from "@/features/invitations/twilioClientProvider";
-import { useConversationUnread } from "@/features/invitations/useConversationUnread";
-import { useInvitationSession } from "@/features/invitations/useInvitationSession";
+import { TwilioChatProvider } from "@/features/invitations/TwilioChatProvider";
+import { useConversationUnread } from "@/features/invitations/hooks/useConversationUnread";
+import { useInvitationSession } from "@/features/invitations/hooks/useInvitationSession";
 import { PrintOnlyLetter } from "@/features/templates/PrintOnlyLetter";
 import { ScaledLetterPreview } from "@/features/templates/ScaledLetterPreview";
 import { useSpeakerInvitation } from "@/features/templates/useSpeakerInvitation";
