+++ src/features/invitations/hooks/useFirstUnreadIndex.ts
