+++ src/features/invitations/utils/statusStripBg.ts
