+++ src/features/plan-speakers/SpeakerActionPicker.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import type { Speaker } from "@/lib/types";
 import type { WithId } from "@/hooks/_sub";
-import { useLatestInvitation } from "@/features/invitations/useLatestInvitation";
+import { useLatestInvitation } from "@/features/invitations/hooks/useLatestInvitation";
 import { isPlausiblePhone } from "@/features/templates/smsInvitation";
 import { isValidEmail } from "@/lib/email";
 import { MissingContactPrompt } from "./MissingContactPrompt";
