+++ src/features/schedule/SundayCardHeader.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router";
 import type { MeetingType, NonMeetingSunday } from "@/lib/types";
 import { cn } from "@/lib/cn";
-import { useSundayInvitationsSummary } from "@/features/invitations/useSundayInvitationsSummary";
+import { useSundayInvitationsSummary } from "@/features/invitations/hooks/useSundayInvitationsSummary";
 import type { KindVariant } from "./kindLabel";
 import { formatShortDate, formatCountdown } from "./dateFormat";
 import { SundayTypeMenu } from "./SundayTypeMenu";
